requirements.txt
click>=8.1.3
setuptools>=60.2.0
pandas>=1.5.0
numpy>=1.23.3
fastparquet>=0.8.3
pyarrow
wget>=3.2
yaspin
PyYAML
gensim>4
scikit-learn
tensorflow
